Changes related to solution of the real problem,
1. Take request with monthly inventory and campaign information as Json.
2. Apply the values to the 1/0 Knapsack algorithm to get high throughput
3. Bag the items which are picked for getting the high throughput
4. Set values to the response bean and send Json response.

+++ src/main/java/se/invidi/campaignbuilder/model/CampaignInformationPerCustomer.java
@@ -4,8 +4,9 @@ import lombok.Data;
 
 @Data
 public class CampaignInformationPerCustomer {
-    private String customer;
-    private short numberOfCampaignsToSell;
-    private long totalImpressions;
-    private int totalRevenue;
+
+  private String customer;
+  private int numberOfCampaignsToSell;
+  private long totalImpressions;
+  private int totalRevenue;
 }
